% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ticks_title.R
\name{add_x_axis}
\alias{add_x_axis}
\alias{add_y_axis}
\title{Add tick marks to x/y axis

A nicer set of default tick marks}
\usage{
add_x_axis(ticks_x, labels = ticks_x, tick = TRUE, lwd = 0, lwd.ticks = 1)

add_y_axis(ticks_y, labels = ticks_y, tick = TRUE, lwd = 0, lwd.ticks = 1)
}
\arguments{
\item{ticks_x}{location of x tick mark}

\item{labels}{tick label}

\item{tick}{Display tick mark}

\item{lwd}{width of line}

\item{lwd.ticks}{width of tick mark}

\item{ticks_y}{location of y tick mark}
}
\description{
Add tick marks to x/y axis

A nicer set of default tick marks
}
