% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot.R
\name{barplot.prettyB}
\alias{barplot.prettyB}
\alias{barplot_p}
\title{PrettyB barplot.default function}
\usage{
\method{barplot}{prettyB}(
  height,
  width = 1,
  space = NULL,
  names.arg = NULL,
  legend.text = NULL,
  beside = FALSE,
  horiz = FALSE,
  density = NULL,
  angle = 45,
  col = NULL,
  border = par("fg"),
  main = NULL,
  sub = NULL,
  xlab = NULL,
  ylab = NULL,
  xlim = NULL,
  ylim = NULL,
  xpd = TRUE,
  log = "",
  axes = TRUE,
  axisnames = TRUE,
  cex.axis = par("cex.axis"),
  cex.names = par("cex.axis"),
  inside = TRUE,
  plot = TRUE,
  axis.lty = 0,
  offset = 0,
  add = FALSE,
  args.legend = NULL,
  ...
)

barplot_p(
  height,
  width = 1,
  space = NULL,
  names.arg = NULL,
  legend.text = NULL,
  beside = FALSE,
  horiz = FALSE,
  density = NULL,
  angle = 45,
  col = NULL,
  border = par("fg"),
  main = NULL,
  sub = NULL,
  xlab = NULL,
  ylab = NULL,
  xlim = NULL,
  ylim = NULL,
  xpd = TRUE,
  log = "",
  axes = TRUE,
  axisnames = TRUE,
  cex.axis = par("cex.axis"),
  cex.names = par("cex.axis"),
  inside = TRUE,
  plot = TRUE,
  axis.lty = 0,
  offset = 0,
  add = FALSE,
  args.legend = NULL,
  ...
)
}
\arguments{
\item{height, width, space}{See \code{?graphics::barplot.default}}

\item{names.arg, legend.text, beside}{See \code{?graphics::barplot.default}}

\item{horiz, density, angle, col}{See \code{?graphics::barplot.default}}

\item{border, main, sub, xlab, ylab}{See \code{?graphics::barplot.default}}

\item{xlim, ylim, xpd, log, axes}{See \code{?graphics::barplot.default}}

\item{axisnames, cex.axis, cex.names}{See \code{?graphics::barplot.default}}

\item{inside, plot, axis.lty}{See \code{?graphics::barplot.default}}

\item{offset, add, args.legend}{See \code{?graphics::barplot.default}}

\item{...}{See \code{?graphics::barplot.default}}
}
\description{
This function overrides the default arguments. See
\code{?graphics::barplot.default}
}
\examples{
barplot(VADeaths)
}
