% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot.R
\name{boxplot.formula}
\alias{boxplot.formula}
\title{PrettyB boxplot.formula function}
\usage{
\method{boxplot}{formula}(formula, data = NULL, ..., subset,
  na.action = NULL, drop = FALSE, sep = ".", lex.order = FALSE)
}
\arguments{
\item{formula, data, ...}{See \code{?graphics::boxplot.formula}}

\item{subset, na.action, drop}{See \code{?graphics::boxplot.formula}}

\item{sep, lex.order}{See \code{?graphics::boxplot.formula}}
}
\description{
This function overrides the default arguments. See
\code{?graphics::boxplot.formula}. This function is identical to
\code{?graphics::boxplot.formula}. However, since
\code{?graphics::boxplot.formula} isn't exported, I had to make a copy.
}
\examples{
boxplot(count ~ spray, data = InsectSprays)
}
