% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hist.R
\name{hist.prettyB}
\alias{hist.prettyB}
\alias{hist_p}
\title{PrettyB hist.default function}
\usage{
\method{hist}{prettyB}(
  x,
  breaks = "Sturges",
  freq = NULL,
  probability = !freq,
  include.lowest = TRUE,
  right = TRUE,
  density = NULL,
  angle = 45,
  col = NULL,
  border = NULL,
  main = paste("Histogram of", xname),
  xlim = range(breaks),
  ylim = NULL,
  xlab = xname,
  ylab,
  axes = TRUE,
  plot = TRUE,
  labels = FALSE,
  nclass = NULL,
  warn.unused = TRUE,
  ...
)

hist_p(
  x,
  breaks = "Sturges",
  freq = NULL,
  probability = !freq,
  include.lowest = TRUE,
  right = TRUE,
  density = NULL,
  angle = 45,
  col = NULL,
  border = NULL,
  main = paste("Histogram of", xname),
  xlim = range(breaks),
  ylim = NULL,
  xlab = xname,
  ylab,
  axes = TRUE,
  plot = TRUE,
  labels = FALSE,
  nclass = NULL,
  warn.unused = TRUE,
  ...
)
}
\arguments{
\item{x, breaks, freq, probability}{See \code{?graphics::hist.default}}

\item{include.lowest, right, density}{See \code{?graphics::hist.default}}

\item{angle, col, border, main}{See \code{?graphics::hist.default}}

\item{xlim, ylim, xlab, ylab, axes, plot}{See \code{?graphics::hist.default}}

\item{labels, nclass, warn.unused}{See \code{?graphics::hist.default}}

\item{...}{See \code{?graphics::hist.default}}
}
\description{
This function overrides the default arguments. See
\code{?graphics::hist.default}
}
\examples{
x = rlnorm(100)
hist(x)
}
