% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot}
\alias{plot}
\alias{plot.default}
\title{PrettyB's plotting function}
\usage{
plot(x, y, ...)

\method{plot}{default}(x, y = NULL, ...)
}
\arguments{
\item{x}{x parameter passed to plot}

\item{y}{y parameter passed to plot}

\item{...}{other parameters passed to plot}
}
\description{
PrettyB's redefined plotting functions
}
