% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prettyB-package.R
\docType{package}
\name{prettyB-package}
\alias{prettyB-package}
\alias{prettyB}
\alias{prettyb}
\title{The prettyB package}
\description{
Anyone who uses R Base graphics, have a 100 and 1 tweaks that they use to make the
figures more presentable. This package aims to capture the tweaks in one place.
By masking_standard plotting functions, we can automatically make base graphics a bit more pretty.
}
\seealso{
\url{https://github.com/jumpingrivers/prettyB}
}
\author{
\email{csgillespie@gmail.com}
}
\keyword{package}
