% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qqnorm.R
\name{qqline}
\alias{qqline}
\title{Replacement for qqline}
\usage{
qqline(y, ...)
}
\arguments{
\item{y}{The second or only data sample.}

\item{...}{Graphical parameters passed to \code{stats::qqline}.}
}
\description{
See ?stats::qqline for details
}
