% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/par.R
\name{reset_prettyB}
\alias{reset_prettyB}
\alias{setup_prettyB}
\title{Set up par and palette

Sets up par and palette}
\usage{
reset_prettyB()

setup_prettyB()
}
\description{
Set up par and palette

Sets up par and palette
}
