% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_set}
\alias{theme_set}
\title{Set the current theme

Set's the current plotting theme.}
\usage{
theme_set(theme = NULL)
}
\arguments{
\item{theme}{Values are "minimal" or "expand"}
}
\description{
Set the current theme

Set's the current plotting theme.
}
