% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.default}
\alias{plot.default}
\title{PrettyB plot.default function}
\usage{
\method{plot}{default}(x, y = NULL, type = "p", xlim = NULL,
  ylim = NULL, log = "", main = NULL, sub = NULL, xlab = NULL,
  ylab = NULL, ann = par("ann"), axes = TRUE, frame.plot = axes,
  panel.first = NULL, panel.last = NULL, asp = NA, xgap.axis = NA,
  ygap.axis = NA, ...)
}
\arguments{
\item{x, y, type, log}{See \code{?graphics::plot.default}}

\item{xlim, ylim}{See \code{?graphics::plot.default}}

\item{main, sub, xlab, ylab}{See \code{?graphics::plot.default}}

\item{ann, axes}{See \code{?graphics::plot.default}}

\item{frame.plot, panel.first, panel.last}{See \code{?graphics::plot.default}}

\item{asp, xgap.axis, ygap.axis}{See \code{?graphics::plot.default}}

\item{...}{See \code{?graphics::plot.default}}
}
\description{
This function overrides the default arguments. See
\code{?graphics::plot.default}
}
