% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot}
\alias{plot}
\alias{plot.default}
\title{PrettyB's plotting function}
\usage{
plot(x, y, ...)

\method{plot}{default}(x, y = NULL, ...)
}
\arguments{
\item{x}{x parameter passed to plot}

\item{y}{y parameter passed to plot}

\item{...}{other parameters passed to plot}
}
\description{
PrettyB's redefined plotting functions
}
\examples{
# PrettyB overloads standard plotting functions
plot(1:10)
# All arguments still work as expect
plot(iris$Sepal.Length, iris$Sepal.Width, bg = "red")
# Use par as usual
op = par(mfrow = c(1, 2))
plot(rnorm(10), xlab = "X")
plot(rnorm(10), main = "A title")
par(op)
}
